+++ netaio/client.py
@@ -1,4 +1,5 @@
 from .auth import AuthPluginProtocol
+from .cipher import CipherPluginProtocol
 from .common import (
     Header,
     AuthFields,
@@ -11,7 +12,6 @@ from .common import (
     Handler,
     default_client_logger
 )
-from .encryption import EncryptionPluginProtocol
 from typing import Callable, Coroutine, Hashable
 import asyncio
 import logging
@@ -24,11 +24,11 @@ class TCPClient:
     header_class: type[HeaderProtocol]
     body_class: type[BodyProtocol]
     message_class: type[MessageProtocol]
-    handlers: dict[Hashable, tuple[Handler, AuthPluginProtocol|None, EncryptionPluginProtocol|None]]
+    handlers: dict[Hashable, tuple[Handler, AuthPluginProtocol|None, CipherPluginProtocol|None]]
     extract_keys: Callable[[MessageProtocol], list[Hashable]]
     logger: logging.Logger
     auth_plugin: AuthPluginProtocol
-    encrypt_plugin: EncryptionPluginProtocol
+    cipher_plugin: CipherPluginProtocol
 
     def __init__(
             self, host: str = "127.0.0.1", port: int = 8888,
@@ -39,7 +39,7 @@ class TCPClient:
             extract_keys: Callable[[MessageProtocol], list[Hashable]] = keys_extractor,
             logger: logging.Logger = default_client_logger,
             auth_plugin: AuthPluginProtocol = None,
-            encrypt_plugin: EncryptionPluginProtocol = None
+            cipher_plugin: CipherPluginProtocol = None
         ):
         """Initialize the TCPClient.
 
@@ -53,7 +53,7 @@ class TCPClient:
             extract_keys: A function that extracts the keys from a message.
             logger: The logger to use.
             auth_plugin: The auth plugin to use.
-            encrypt_plugin: The encryption plugin to use.
+            cipher_plugin: The cipher plugin to use.
         """
         self.hosts = {}
         self.default_host = (host, port)
@@ -65,43 +65,43 @@ class TCPClient:
         self.extract_keys = extract_keys
         self.logger = logger
         self.auth_plugin = auth_plugin
-        self.encrypt_plugin = encrypt_plugin
+        self.cipher_plugin = cipher_plugin
 
     def add_handler(
             self, key: Hashable,
             handler: Handler,
             auth_plugin: AuthPluginProtocol = None,
-            encrypt_plugin: EncryptionPluginProtocol = None
+            cipher_plugin: CipherPluginProtocol = None
         ):
         """Register a handler for a specific key. The handler must
             accept a MessageProtocol object as an argument and return
             MessageProtocol, None, or a Coroutine that resolves to
             MessageProtocol | None. If an auth plugin is provided, it
             will be used to check the message in addition to any auth
-            plugin that is set on the client. If an encrypt plugin is
+            plugin that is set on the client. If a cipher plugin is
             provided, it will be used to decrypt the message in addition
-            to any encryption plugin that is set on the client.
+            to any cipher plugin that is set on the client.
         """
         self.logger.debug("Adding handler for key=%s", key)
-        self.handlers[key] = (handler, auth_plugin, encrypt_plugin)
+        self.handlers[key] = (handler, auth_plugin, cipher_plugin)
 
     def on(
             self, key: Hashable,
             auth_plugin: AuthPluginProtocol = None,
-            encrypt_plugin: EncryptionPluginProtocol = None
+            cipher_plugin: CipherPluginProtocol = None
         ):
         """Decorator to register a handler for a specific key. The
             handler must accept a MessageProtocol object as an argument
             and return a MessageProtocol, None, or a Coroutine that
             resolves to a MessageProtocol or None. If an auth plugin is
             provided, it will be used to check the message in addition
-            to any auth plugin that is set on the client. If an encrypt
+            to any auth plugin that is set on the client. If a cipher
             plugin is provided, it will be used to decrypt the message
-            in addition to any encryption plugin that is set on the
+            in addition to any cipher plugin that is set on the
             client.
         """
         def decorator(func: Handler):
-            self.add_handler(key, func, auth_plugin, encrypt_plugin)
+            self.add_handler(key, func, auth_plugin, cipher_plugin)
             return func
         return decorator
 
@@ -115,38 +115,38 @@ class TCPClient:
 
     async def send(
             self, message: MessageProtocol, server: tuple[str, int] = None,
-            use_auth: bool = True, use_encryption: bool = True,
+            use_auth: bool = True, use_cipher: bool = True,
             auth_plugin: AuthPluginProtocol|None = None,
-            encrypt_plugin: EncryptionPluginProtocol|None = None
+            cipher_plugin: CipherPluginProtocol|None = None
         ):
         """Send a message to the server. If use_auth is True and an auth
             plugin is set, it will be called to set the auth fields on
             the message. If an auth plugin is provided, it will be used
             to authorize the message in addition to any auth plugin that
-            is set on the client. If an encrypt plugin is provided, it
+            is set on the client. If a cipher plugin is provided, it
             will be used to encrypt the message in addition to any
-            encryption plugin that is set on the client. If use_auth is
+            cipher plugin that is set on the client. If use_auth is
             False, the auth plugin set on the client will not be used.
-            If use_encryption is False, the encryption plugin set on the
+            If use_cipher is False, the cipher plugin set on the
             client will not be used.
         """
         server = server or self.default_host
 
 
-        # inner encryption
-        if encrypt_plugin is not None:
-            self.logger.debug("Calling encrypt_plugin.encrypt on message")
-            message = encrypt_plugin.encrypt(message)
+        # inner cipher
+        if cipher_plugin is not None:
+            self.logger.debug("Calling cipher_plugin.encrypt on message")
+            message = cipher_plugin.encrypt(message)
 
         # inner auth
         if auth_plugin is not None:
             self.logger.debug("Calling auth_plugin.make on auth_data and body")
             auth_plugin.make(message.auth_data, message.body)
 
-        # outer encryption
-        if use_encryption and self.encrypt_plugin is not None:
-            self.logger.debug("Calling self.encrypt_plugin.encrypt on message")
-            message = self.encrypt_plugin.encrypt(message)
+        # outer cipher
+        if use_cipher and self.cipher_plugin is not None:
+            self.logger.debug("Calling self.cipher_plugin.encrypt on message")
+            message = self.cipher_plugin.encrypt(message)
 
         # outer auth
         if use_auth and self.auth_plugin is not None:
@@ -161,8 +161,8 @@ class TCPClient:
 
     async def receive_once(
             self, server: tuple[str, int] = None, use_auth: bool = True,
-            use_encryption: bool = True, auth_plugin: AuthPluginProtocol|None = None,
-            encrypt_plugin: EncryptionPluginProtocol|None = None
+            use_cipher: bool = True, auth_plugin: AuthPluginProtocol|None = None,
+            cipher_plugin: CipherPluginProtocol|None = None
         ) -> MessageProtocol|None:
         """Receive a message from the server. If a handler was
             registered for the message key, the handler will be called
@@ -174,12 +174,12 @@ class TCPClient:
             is called, and if the check fails, the message will be
             discarded and None will be returned. If use_auth is False,
             the auth plugin set on the client will not be used. If
-            use_encryption is False, the encryption plugin set on the
+            use_cipher is False, the cipher plugin set on the
             client will not be used. If an auth plugin is provided, it
             will be used to check the message in addition to any auth
-            plugin that is set on the client. If an encrypt plugin is
+            plugin that is set on the client. If a cipher plugin is
             provided, it will be used to decrypt the message in addition
-            to any encryption plugin that is set on the client.
+            to any cipher plugin that is set on the client.
         """
         self.logger.debug("Receiving message from server...")
         server = server or self.default_host
@@ -207,10 +207,10 @@ class TCPClient:
                 self.logger.warning("Message auth failed")
                 return None
 
-        # outer encryption
-        if use_encryption and self.encrypt_plugin is not None:
-            self.logger.debug("Calling encrypt_plugin.decrypt on message")
-            msg = self.encrypt_plugin.decrypt(msg)
+        # outer cipher
+        if use_cipher and self.cipher_plugin is not None:
+            self.logger.debug("Calling cipher_plugin.decrypt on message")
+            msg = self.cipher_plugin.decrypt(msg)
 
         # inner auth
         if auth_plugin is not None:
@@ -219,10 +219,10 @@ class TCPClient:
                 self.logger.warning("Message auth failed")
                 return None
 
-        # inner encryption
-        if encrypt_plugin is not None:
-            self.logger.debug("Calling encrypt_plugin.decrypt on message")
-            msg = encrypt_plugin.decrypt(msg)
+        # inner cipher
+        if cipher_plugin is not None:
+            self.logger.debug("Calling cipher_plugin.decrypt on message")
+            msg = cipher_plugin.decrypt(msg)
 
         keys = self.extract_keys(msg)
         result = None
@@ -230,16 +230,16 @@ class TCPClient:
         self.logger.debug("Message received from server")
         for key in keys:
             if key in self.handlers:
-                handler, auth_plugin, encrypt_plugin = self.handlers[key]
+                handler, auth_plugin, cipher_plugin = self.handlers[key]
 
                 if auth_plugin is not None:
                     if not auth_plugin.check(msg.auth_data, msg.body):
                         self.logger.warning("Message auth failed")
                         return None
 
-                if encrypt_plugin is not None:
-                    self.logger.debug("Calling encrypt_plugin.decrypt on message")
-                    msg = encrypt_plugin.decrypt(msg)
+                if cipher_plugin is not None:
+                    self.logger.debug("Calling cipher_plugin.decrypt on message")
+                    msg = cipher_plugin.decrypt(msg)
 
                 self.logger.debug("Calling handler for key=%s", key)
                 result = handler(msg, writer)
@@ -254,24 +254,24 @@ class TCPClient:
 
     async def receive_loop(
             self, server: tuple[str, int] = None, use_auth: bool = True,
-            use_encryption: bool = True, auth_plugin: AuthPluginProtocol|None = None,
-            encrypt_plugin: EncryptionPluginProtocol|None = None
+            use_cipher: bool = True, auth_plugin: AuthPluginProtocol|None = None,
+            cipher_plugin: CipherPluginProtocol|None = None
         ):
         """Receive messages from the server indefinitely. Use with
             asyncio.create_task() to run concurrently, then use
             task.cancel() to stop. If use_auth is False, the auth plugin
-            set on the client will not be used. If use_encryption is
-            False, the encryption plugin set on the client will not be
+            set on the client will not be used. If use_cipher is
+            False, the cipher plugin set on the client will not be
             used. If an auth plugin is provided, it will be used to
             check the message in addition to any auth plugin that is set
-            on the client. If an encrypt plugin is provided, it will be
-            used to decrypt the message in addition to any encryption
+            on the client. If a cipher plugin is provided, it will be
+            used to decrypt the message in addition to any cipher
             plugin that is set on the client.
         """
         server = server or self.default_host
         while True:
             try:
-                await self.receive_once(server, use_auth, use_encryption, auth_plugin, encrypt_plugin)
+                await self.receive_once(server, use_auth, use_cipher, auth_plugin, cipher_plugin)
             except asyncio.CancelledError:
                 self.logger.info("Receive loop cancelled")
                 break