+++ netaio/server.py
@@ -1,4 +1,5 @@
 from .auth import AuthPluginProtocol
+from .cipher import CipherPluginProtocol
 from .common import (
     Header,
     AuthFields,
@@ -12,7 +13,6 @@ from .common import (
     Handler,
     default_server_logger
 )
-from .encryption import EncryptionPluginProtocol
 from typing import Callable, Coroutine, Hashable
 import asyncio
 import logging
@@ -25,7 +25,7 @@ def not_found_handler(*_) -> MessageProtocol | None:
 class TCPServer:
     host: str
     port: int
-    handlers: dict[Hashable, tuple[Handler, AuthPluginProtocol|None, EncryptionPluginProtocol|None]]
+    handlers: dict[Hashable, tuple[Handler, AuthPluginProtocol|None, CipherPluginProtocol|None]]
     default_handler: Handler
     header_class: type[HeaderProtocol]
     body_class: type[BodyProtocol]
@@ -36,7 +36,7 @@ class TCPServer:
     clients: set[asyncio.StreamWriter]
     logger: logging.Logger
     auth_plugin: AuthPluginProtocol
-    encrypt_plugin: EncryptionPluginProtocol
+    cipher_plugin: CipherPluginProtocol
     def __init__(
             self, host: str = "0.0.0.0", port: int = 8888,
             header_class: type[HeaderProtocol] = Header,
@@ -47,7 +47,7 @@ class TCPServer:
             default_handler: Handler = not_found_handler,
             logger: logging.Logger = default_server_logger,
             auth_plugin: AuthPluginProtocol = None,
-            encrypt_plugin: EncryptionPluginProtocol = None
+            cipher_plugin: CipherPluginProtocol = None
         ):
         """Initialize the TCPServer.
 
@@ -63,7 +63,7 @@ class TCPServer:
                 do not match any registered handler keys.
             logger: The logger to use.
             auth_plugin: The auth plugin to use.
-            encrypt_plugin: The encrypt plugin to use.
+            cipher_plugin: The cipher plugin to use.
         """
         self.host = host
         self.port = port
@@ -78,46 +78,46 @@ class TCPServer:
         self.default_handler = default_handler
         self.logger = logger
         self.auth_plugin = auth_plugin
-        self.encrypt_plugin = encrypt_plugin
+        self.cipher_plugin = cipher_plugin
 
     def add_handler(
             self, key: Hashable,
             handler: Handler,
             auth_plugin: AuthPluginProtocol = None,
-            encrypt_plugin: EncryptionPluginProtocol = None
+            cipher_plugin: CipherPluginProtocol = None
         ):
         """Register a handler for a specific key. The handler must
             accept a MessageProtocol object as an argument and return a
             MessageProtocol, None, or a Coroutine that resolves to
             MessageProtocol | None. If an auth plugin is provided, it
             will be used to check the message in addition to any auth
-            plugin that is set on the server. If an encrypt plugin is
+            plugin that is set on the server. If a cipher plugin is
             provided, it will be used to decrypt the message in addition
-            to any encryption plugin that is set on the server. These
+            to any cipher plugin that is set on the server. These
             plugins will also be used for preparing any response
             message sent by the handler.
         """
         self.logger.debug("Adding handler for key=%s", key)
-        self.handlers[key] = (handler, auth_plugin, encrypt_plugin)
+        self.handlers[key] = (handler, auth_plugin, cipher_plugin)
 
     def on(
             self, key: Hashable,
             auth_plugin: AuthPluginProtocol = None,
-            encrypt_plugin: EncryptionPluginProtocol = None
+            cipher_plugin: CipherPluginProtocol = None
         ):
         """Decorator to register a handler for a specific key. The
             handler must accept a MessageProtocol object as an argument
             and return a MessageProtocol, None, or a Coroutine that
             resolves to a MessageProtocol or None. If an auth plugin is
             provided, it will be used to check the message in addition
-            to any auth plugin that is set on the server. If an encrypt
+            to any auth plugin that is set on the server. If a cipher
             plugin is provided, it will be used to decrypt the message in
-            addition to any encryption plugin that is set on the server.
+            addition to any cipher plugin that is set on the server.
             These plugins will also be used for preparing any response
             message sent by the handler.
         """
         def decorator(func: Handler):
-            self.add_handler(key, func, auth_plugin, encrypt_plugin)
+            self.add_handler(key, func, auth_plugin, cipher_plugin)
             return func
         return decorator
 
@@ -143,14 +143,14 @@ class TCPServer:
 
     async def handle_client(
             self, reader: asyncio.StreamReader, writer: asyncio.StreamWriter,
-            use_auth: bool = True, use_encryption: bool = True
+            use_auth: bool = True, use_cipher: bool = True
         ):
         """Handle a client connection. When a client connects, it is
             added to the clients set. The client is then read from until
             the connection is lost, and the proper handlers are called
             if they are defined and the message is valid. If use_auth is
             False, the auth plugin set on the server will not be used.
-            If use_encryption is False, the encryption plugin set on the
+            If use_cipher is False, the cipher plugin set on the
             server will not be used.
         """
         self.logger.info("Client connected from %s", writer.get_extra_info("peername"))
@@ -160,7 +160,7 @@ class TCPServer:
         try:
             while True:
                 auth_plugin = None
-                encrypt_plugin = None
+                cipher_plugin = None
                 header_bytes = await reader.readexactly(header_length)
                 header = self.header_class.decode(header_bytes)
 
@@ -186,22 +186,22 @@ class TCPServer:
                         if not self.auth_plugin.check(message.auth_data, message.body):
                             self.logger.warning("Invalid auth_fields received from %s", writer.get_extra_info("peername"))
                             response = self.auth_plugin.error()
-                            await self.send(writer, response, use_auth=False, use_encryption=False)
+                            await self.send(writer, response, use_auth=False, use_cipher=False)
                             continue
                         else:
                             self.logger.debug("Valid auth_fields received from %s", writer.get_extra_info("peername"))
 
-                    # outer encryption
-                    if use_encryption and self.encrypt_plugin is not None:
-                        self.logger.debug("Calling self.encrypt_plugin.decrypt on message")
-                        message = self.encrypt_plugin.decrypt(message)
+                    # outer cipher
+                    if use_cipher and self.cipher_plugin is not None:
+                        self.logger.debug("Calling self.cipher_plugin.decrypt on message")
+                        message = self.cipher_plugin.decrypt(message)
 
                     keys = self.extract_keys(message)
                     self.logger.debug("Message received from %s with keys=%s", writer.get_extra_info("peername"), keys)
 
                     for key in keys:
                         if key in self.handlers:
-                            handler, auth_plugin, encrypt_plugin = self.handlers[key]
+                            handler, auth_plugin, cipher_plugin = self.handlers[key]
 
                             # inner auth
                             if auth_plugin is not None:
@@ -211,10 +211,10 @@ class TCPServer:
                                     response = auth_plugin.error()
                                     break
 
-                            # inner encryption
-                            if encrypt_plugin is not None:
-                                self.logger.debug("Calling encrypt_plugin.decrypt on message")
-                                message = encrypt_plugin.decrypt(message)
+                            # inner cipher
+                            if cipher_plugin is not None:
+                                self.logger.debug("Calling cipher_plugin.decrypt on message")
+                                message = cipher_plugin.decrypt(message)
 
                             self.logger.debug("Calling handler for key=%s", key)
                             response = handler(message, writer)
@@ -226,27 +226,27 @@ class TCPServer:
                         response = self.default_handler(message, writer)
 
                 if response is not None:
-                    # inner encryption
-                    if encrypt_plugin is not None:
-                        self.logger.debug("Calling encrypt_plugin.encrypt on response")
-                        response = encrypt_plugin.encrypt(response)
+                    # inner cipher
+                    if cipher_plugin is not None:
+                        self.logger.debug("Calling cipher_plugin.encrypt on response")
+                        response = cipher_plugin.encrypt(response)
 
                     # inner auth
                     if auth_plugin is not None:
                         self.logger.debug("Calling auth_plugin.make on response.body (handler)")
                         auth_plugin.make(response.auth_data, response.body)
 
-                    # outer encryption
-                    if use_encryption and self.encrypt_plugin is not None:
-                        self.logger.debug("Calling encrypt_plugin.encrypt on response")
-                        response = self.encrypt_plugin.encrypt(response)
+                    # outer cipher
+                    if use_cipher and self.cipher_plugin is not None:
+                        self.logger.debug("Calling cipher_plugin.encrypt on response")
+                        response = self.cipher_plugin.encrypt(response)
 
                     # outer auth
                     if use_auth and self.auth_plugin is not None:
                         self.logger.debug("Calling self.auth_plugin.make on response.body")
                         self.auth_plugin.make(response.auth_data, response.body)
 
-                    await self.send(writer, response, use_auth=False, use_encryption=False)
+                    await self.send(writer, response, use_auth=False, use_cipher=False)
         except asyncio.IncompleteReadError:
             self.logger.info("Client disconnected from %s", writer.get_extra_info("peername"))
             pass  # Client disconnected
@@ -266,10 +266,10 @@ class TCPServer:
             writer.close()
             await writer.wait_closed()
 
-    async def start(self, use_auth: bool = True, use_encryption: bool = True):
+    async def start(self, use_auth: bool = True, use_cipher: bool = True):
         """Start the server."""
         server = await asyncio.start_server(
-            lambda r, w: self.handle_client(r, w, use_auth, use_encryption),
+            lambda r, w: self.handle_client(r, w, use_auth, use_cipher),
             self.host, self.port
         )
         async with server:
@@ -279,34 +279,34 @@ class TCPServer:
     async def send(
             self, client: asyncio.StreamWriter, message: MessageProtocol,
             collection: set = None, use_auth: bool = True,
-            use_encryption: bool = True, auth_plugin: AuthPluginProtocol|None = None,
-            encrypt_plugin: EncryptionPluginProtocol|None = None
+            use_cipher: bool = True, auth_plugin: AuthPluginProtocol|None = None,
+            cipher_plugin: CipherPluginProtocol|None = None
         ):
         """Helper coroutine to send a message to a client. On error, it
             logs the exception and removes the client from the given
             collection. If an auth plugin is provided, it will be used
             to authorize the message in addition to any auth plugin that
-            is set on the server. If an encrypt plugin is provided, it
+            is set on the server. If a cipher plugin is provided, it
             will be used to encrypt the message in addition to any
-            encryption plugin that is set on the server. If use_auth is
+            cipher plugin that is set on the server. If use_auth is
             False, the auth plugin set on the server will not be used.
-            If use_encryption is False, the encryption plugin set on the
+            If use_cipher is False, the cipher plugin set on the
             server will not be used.
         """
-        # inner encryption
-        if encrypt_plugin is not None:
-            self.logger.debug("Calling encrypt_plugin.encrypt on message")
-            message = encrypt_plugin.encrypt(message)
+        # inner cipher
+        if cipher_plugin is not None:
+            self.logger.debug("Calling cipher_plugin.encrypt on message")
+            message = cipher_plugin.encrypt(message)
 
         # inner auth
         if auth_plugin is not None:
             self.logger.debug("Calling auth_plugin.make on auth_data and body")
             auth_plugin.make(message.auth_data, message.body)
 
-        # outer encryption
-        if use_encryption and self.encrypt_plugin is not None:
-            self.logger.debug("Calling self.encrypt_plugin.encrypt on message")
-            message = self.encrypt_plugin.encrypt(message)
+        # outer cipher
+        if use_cipher and self.cipher_plugin is not None:
+            self.logger.debug("Calling self.cipher_plugin.encrypt on message")
+            message = self.cipher_plugin.encrypt(message)
 
         # outer auth
         if use_auth and self.auth_plugin is not None:
@@ -325,35 +325,35 @@ class TCPServer:
 
     async def broadcast(
             self, message: MessageProtocol, use_auth: bool = True,
-            use_encryption: bool = True, auth_plugin: AuthPluginProtocol|None = None,
-            encrypt_plugin: EncryptionPluginProtocol|None = None
+            use_cipher: bool = True, auth_plugin: AuthPluginProtocol|None = None,
+            cipher_plugin: CipherPluginProtocol|None = None
         ):
         """Send the message to all connected clients concurrently using
             asyncio.gather. If an auth plugin is provided, it will be
             used to authorize the message in addition to any auth plugin
-            that is set on the server. If an encrypt plugin is provided,
+            that is set on the server. If a cipher plugin is provided,
             it will be used to encrypt the message in addition to any
-            encryption plugin that is set on the server. If use_auth is
+            cipher plugin that is set on the server. If use_auth is
             False, the auth plugin set on the server will not be used. If
-            use_encryption is False, the encryption plugin set on the
+            use_cipher is False, the cipher plugin set on the
             server will not be used.
         """
         self.logger.debug("Broadcasting message to all clients")
 
-        # inner encryption
-        if use_encryption and encrypt_plugin is not None:
-            self.logger.debug("Calling encrypt_plugin.encrypt on message")
-            message = encrypt_plugin.encrypt(message)
+        # inner cipher
+        if use_cipher and cipher_plugin is not None:
+            self.logger.debug("Calling cipher_plugin.encrypt on message")
+            message = cipher_plugin.encrypt(message)
 
         # inner auth
         if use_auth and auth_plugin is not None:
             self.logger.debug("Calling auth_plugin.make on message.body (broadcast)")
             auth_plugin.make(message.auth_data, message.body)
 
-        # outer encryption
-        if use_encryption and self.encrypt_plugin is not None:
-            self.logger.debug("Calling encrypt_plugin.encrypt on message")
-            message = self.encrypt_plugin.encrypt(message)
+        # outer cipher
+        if use_cipher and self.cipher_plugin is not None:
+            self.logger.debug("Calling cipher_plugin.encrypt on message")
+            message = self.cipher_plugin.encrypt(message)
 
         # outer auth
         if use_auth and self.auth_plugin is not None:
@@ -365,18 +365,18 @@ class TCPServer:
 
     async def notify(
             self, key: Hashable, message: MessageProtocol, use_auth: bool = True,
-            use_encryption: bool = True, auth_plugin: AuthPluginProtocol|None = None,
-            encrypt_plugin: EncryptionPluginProtocol|None = None
+            use_cipher: bool = True, auth_plugin: AuthPluginProtocol|None = None,
+            cipher_plugin: CipherPluginProtocol|None = None
         ):
         """Send the message to all subscribed clients for the given key
             concurrently using asyncio.gather. If an auth plugin is
             provided, it will be used to authorize the message in
             addition to any auth plugin that is set on the server. If an
-            encrypt plugin is provided, it will be used to encrypt the
-            message in addition to any encryption plugin that is set on
+            cipher plugin is provided, it will be used to encrypt the
+            message in addition to any cipher plugin that is set on
             the server. If use_auth is False, the auth plugin set on the
-            server will not be used. If use_encryption is False, the
-            encryption plugin set on the server will not be used.
+            server will not be used. If use_cipher is False, the
+            cipher plugin set on the server will not be used.
         """
         if key not in self.subscriptions:
             self.logger.debug("No subscribers found for key=%s, skipping notification", key)
@@ -384,20 +384,20 @@ class TCPServer:
 
         self.logger.debug("Notifying %d clients for key=%s", len(self.subscriptions[key]), key)
 
-        # inner encryption
-        if use_encryption and encrypt_plugin is not None:
-            self.logger.debug("Calling encrypt_plugin.encrypt on message")
-            message = encrypt_plugin.encrypt(message)
+        # inner cipher
+        if use_cipher and cipher_plugin is not None:
+            self.logger.debug("Calling cipher_plugin.encrypt on message")
+            message = cipher_plugin.encrypt(message)
 
         # inner auth
         if use_auth and auth_plugin is not None:
             self.logger.debug("Calling auth_plugin.make on message.body (notify)")
             auth_plugin.make(message.auth_data, message.body)
 
-        # outer encryption
-        if use_encryption and self.encrypt_plugin is not None:
-            self.logger.debug("Calling encrypt_plugin.encrypt on message")
-            message = self.encrypt_plugin.encrypt(message)
+        # outer cipher
+        if use_cipher and self.cipher_plugin is not None:
+            self.logger.debug("Calling cipher_plugin.encrypt on message")
+            message = self.cipher_plugin.encrypt(message)
 
         # outer auth
         if use_auth and self.auth_plugin is not None: